+++ include/format.hpp
@@ -4,7 +4,7 @@
 #include <string>
 #include <SFML/System/Time.hpp>
 
-std::string formatTime(double seconds);
+std::string formatTime(float seconds);
 std::string formatTime(const sf::Time& time);
 
 #endif // FORMAT_HPP
